requirements.txt
pillow>=10.0.0
opencv-python-headless>=4.8.0.74
numpy>=1.25.0
gradio>=3.39.0
ultralytics>=8.0.0
deep-translator>=1.11.0
transformers>=4.35.0
translators>=5.4.0
